refactor/organization-enroll: consolidate organization settings into a single LiveView and rename organization routes
